refactor: 기존 fitness 패키지 분리
- fitness 서비스, 리포지터리 관련 패키지 분리
