+++ src/test/java/com/flab/eattofit/exercise/application/fitness/FitnessServiceTest.java
@@ -1,6 +1,7 @@
 package com.flab.eattofit.exercise.application.fitness;
 
-import com.flab.eattofit.exercise.application.fitness.dto.FitnessCreateRequest;
+import com.flab.eattofit.exercise.application.fitness.fitness.FitnessService;
+import com.flab.eattofit.exercise.application.fitness.fitness.dto.FitnessCreateRequest;
 import com.flab.eattofit.exercise.domain.fitness.fitness.Fitness;
 import com.flab.eattofit.exercise.domain.fitness.fitness.FitnessRepository;
 import com.flab.eattofit.exercise.exception.fitness.exceptions.FitnessAlreadyExistedException;
