+++ src/test/java/com/flab/eattofit/helper/MockBeanInjection.java
@@ -2,7 +2,7 @@ package com.flab.eattofit.helper;
 
 import com.amazonaws.services.s3.AmazonS3;
 import com.fasterxml.jackson.databind.json.JsonMapper;
-import com.flab.eattofit.exercise.application.fitness.FitnessService;
+import com.flab.eattofit.exercise.application.fitness.fitness.FitnessService;
 import com.flab.eattofit.exercise.application.sports.SportsService;
 import com.flab.eattofit.exercise.domain.fitness.fitness.FitnessRepository;
 import com.flab.eattofit.exercise.domain.sports.SportsRepository;
